requirements.txt
scikit-learn
kafka-python
pyspark==3.5.0